PAAC-130: Updated calculator factory object to select correct calculator object based on find/issupported rather than match with literals.
