+++ app/logic/PensionAllowanceCalculator.scala
@@ -20,49 +20,15 @@ import models._
 
 trait Calculator {
   def summary(previousPeriods:Seq[SummaryResult], contribution:Contribution) : Option[SummaryResult]
+  def isSupported(contribution:Contribution):Boolean
 }
 
 object CalculatorFactory {
-  def get(contribution:Contribution) : Option[Calculator] = contribution match {
-    case Contribution(TaxPeriod(year, _, _ ), _, _) if year < 2014 && year > 2007 => Some(Pre2014Calculator)
-    /*case Contribution(TaxPeriod(year, _, _ ), _, _) if year == 2014 => Some(Year2014Calculator)
-    case Contribution(TaxPeriod(year, _, _ ), _, _) if year == 2015 => Some(Year2015Period1Calculator)
-    case Contribution(TaxPeriod(year, _, _ ), _, _) if year == 2015 => Some(Year2015Period2Calculator)
-    case Contribution(TaxPeriod(year, _, _ ), _, _) if year > 2015 => Some(Post2015Period2Calculator)*/
-    case _ => None
-  }
+  val calculators : List[Calculator] = List(Pre2014Calculator)
+
+  def get(contribution:Contribution) : Option[Calculator] = 
+    calculators.find(_.isSupported(contribution))
 }
-
-object Pre2014Calculator extends Calculator {
-  def summary(previousPeriods:Seq[SummaryResult], contribution: models.Contribution): Option[SummaryResult] = contribution match {
-    case Contribution(TaxPeriod(year, _, _ ), _, _) if year < 2014 && year > 2007 =>
-      val annualAllowance: Long = 50000*100
-      val exceedingAAAmount: Long = (contribution.amounts.definedBenefit - annualAllowance).max(0)
-      val unusedAllowance: Long = (annualAllowance - contribution.amounts.definedBenefit).max(0)
-      val availableAAWithCF: Long = annualAllowance + previousPeriods.slice(0,3).foldLeft(0L)(_+_.unusedAllowance)
-      val availableAAWithCCF: Long = (annualAllowance + previousPeriods.slice(0,2).foldLeft(0L)(_+_.unusedAllowance) - (contribution.amounts.definedBenefit-exceedingAAAmount)).max(0)
-      val chargableAmount: Long = if (previousPeriods.slice(0,3).size < 3) 0 else (contribution.amounts.definedBenefit - availableAAWithCF).max(0)
-
-      Some(SummaryResult(chargableAmount, exceedingAAAmount, annualAllowance, unusedAllowance, availableAAWithCF, availableAAWithCCF))
-    case _ => None
-  }
-}
-
-/*object Year2014Calculator extends Calculator {
-  def summary(contribution: models.Contribution): Option[SummaryResult] = Some(SummaryResult())
-}
-
-object Year2015Period1Calculator extends Calculator {
-  def summary(contribution: models.Contribution): Option[SummaryResult] = Some(SummaryResult())
-}
-
-object Year2015Period2Calculator extends Calculator {
-  def summary(contribution: models.Contribution): Option[SummaryResult] = Some(SummaryResult())
-}
-
-object Post2015Period2Calculator extends Calculator {
-  def summary(contribution: models.Contribution): Option[SummaryResult] = Some(SummaryResult())
-}*/
 
 trait PensionAllowanceCalculator {
   def calculateAllowances(contributions : Seq[Contribution]) : Seq[TaxYearResults] = {
